@@ -8,62 +8,72 @@
     "build:prod": "ng build --configuration=production",
     "test": "ng test",
     "lint": "ng lint",
-    "lint:fix": "ng lint --fix",
-    "e2e": "ng e2e"
+    "lint:fix": "ng lint --fix"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~11.0.5",
-    "@angular/common": "~11.0.5",
-    "@angular/compiler": "~11.0.5",
-    "@angular/core": "~11.0.5",
-    "@angular/forms": "~11.0.5",
-    "@angular/platform-browser": "~11.0.5",
-    "@angular/platform-browser-dynamic": "~11.0.5",
-    "@angular/router": "~11.0.5",
+    "@angular/animations": "~13.1.1",
+    "@angular/cdk": "^13.1.1",
+    "@angular/common": "~13.1.1",
+    "@angular/compiler": "~13.1.1",
+    "@angular/core": "~13.1.1",
+    "@angular/forms": "~13.1.1",
+    "@angular/platform-browser": "~13.1.1",
+    "@angular/platform-browser-dynamic": "~13.1.1",
+    "@angular/router": "~13.1.1",
     "@mapbox/mapbox-gl-geocoder": "^4.7.4",
     "@ngx-translate/core": "^13.0.0",
     "@ngx-translate/http-loader": "^6.0.0",
-    "@taiga-ui/addon-charts": "^2.27.1",
-    "@taiga-ui/addon-commerce": "^2.27.1",
-    "@taiga-ui/addon-doc": "^2.27.1",
-    "@taiga-ui/addon-editor": "^2.27.1",
-    "@taiga-ui/addon-mobile": "^2.27.1",
-    "@taiga-ui/addon-table": "^2.27.1",
-    "@taiga-ui/addon-tablebars": "^2.27.1",
-    "@taiga-ui/cdk": "^2.27.1",
-    "@taiga-ui/core": "^2.27.1",
-    "@taiga-ui/icons": "^2.27.1",
-    "@taiga-ui/kit": "^2.27.1",
+    "@ngxs/store": "^3.7.3",
+    "@taiga-ui/addon-charts": "^2.29.0",
+    "@taiga-ui/addon-commerce": "^2.29.0",
+    "@taiga-ui/addon-doc": "^2.29.0",
+    "@taiga-ui/addon-editor": "^2.29.0",
+    "@taiga-ui/addon-mobile": "^2.29.0",
+    "@taiga-ui/addon-preview": "^2.29.0",
+    "@taiga-ui/addon-table": "^2.29.0",
+    "@taiga-ui/addon-tablebars": "^2.29.0",
+    "@taiga-ui/cdk": "^2.29.0",
+    "@taiga-ui/core": "^2.29.0",
+    "@taiga-ui/icons": "^2.29.0",
+    "@taiga-ui/kit": "^2.29.0",
     "@tinkoff/ng-dompurify": "3.0.0",
     "date-fns": "^2.27.0",
     "dompurify": "2.2.9",
     "mapbox-gl": "^2.6.1",
-    "ngx-mapbox-gl": "^6.0.4",
+    "ngx-mapbox-gl": "^7.1.2",
+    "ngx-mapbox-gl-geocoder-control": "^0.0.1",
     "rxjs": "~6.6.0",
     "tslib": "^2.0.0",
-    "zone.js": "~0.10.2"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1100.5",
-    "@angular/cli": "~11.0.5",
-    "@angular/compiler-cli": "~11.0.5",
+    "@angular-devkit/build-angular": "~13.1.2",
+    "@angular-eslint/builder": "13.0.1",
+    "@angular-eslint/eslint-plugin": "13.0.1",
+    "@angular-eslint/eslint-plugin-template": "13.0.1",
+    "@angular-eslint/schematics": "13.0.1",
+    "@angular-eslint/template-parser": "13.0.1",
+    "@angular/cli": "~13.1.2",
+    "@angular/compiler-cli": "~13.1.1",
+    "@ngxs/devtools-plugin": "^3.7.3",
     "@types/dompurify": "2.2.3",
-    "@types/jasmine": "~3.6.0",
+    "@types/jasmine": "~3.8.0",
     "@types/mapbox__mapbox-gl-geocoder": "^4.7.1",
     "@types/mapbox-gl": "^2.6.0",
     "@types/node": "^12.11.1",
-    "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.1.0",
+    "@typescript-eslint/eslint-plugin": "5.3.0",
+    "@typescript-eslint/parser": "5.3.0",
+    "eslint": "^8.2.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "jasmine-core": "~3.8.0",
+    "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "protractor": "~7.0.0",
-    "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "~4.0.2"
+    "karma-jasmine-html-reporter": "~1.7.0",
+    "prettier": "^2.5.1",
+    "typescript": "~4.5.4"
   }
 }
