@@ -1,23 +1,21 @@
-import {Injectable} from '@angular/core';
+import { Injectable } from '@angular/core';
 import {
-  HttpRequest,
-  HttpHandler,
   HttpEvent,
-  HttpInterceptor
+  HttpHandler,
+  HttpInterceptor,
+  HttpRequest,
 } from '@angular/common/http';
-import {Observable} from 'rxjs';
-import {MapboxService} from './mapbox.service';
-
+import { Observable } from 'rxjs';
+import { MapboxService } from './mapbox.service';
 
 @Injectable()
 export class MapboxInterceptor implements HttpInterceptor {
+  constructor(private readonly mapboxService: MapboxService) {}
 
-  constructor(
-    private readonly mapboxService: MapboxService,
-  ) {
-  }
-
-  intercept(request: HttpRequest<unknown>, next: HttpHandler): Observable<HttpEvent<unknown>> {
+  intercept(
+    request: HttpRequest<unknown>,
+    next: HttpHandler,
+  ): Observable<HttpEvent<unknown>> {
     if (!request.url.startsWith(this.mapboxService.urlPrefix)) {
       return next.handle(request);
     }
@@ -25,6 +23,6 @@ export class MapboxInterceptor implements HttpInterceptor {
     const params = {
       access_token: window.env.mapboxPublicApiKey,
     };
-    return next.handle(request.clone({setParams: params}));
+    return next.handle(request.clone({ setParams: params }));
   }
 }
