@@ -8,25 +8,25 @@ export interface ZmanimResponseDto {
     coordinates?: string[];
     elevation?: number;
     fast_name?: string;
-    yomtov_name?: string
+    yomtov_name?: string;
   };
-  alos?: string;
-  sunrise?: string;
-  misheyakir_10_2?: string;
-  sof_zman_shema_ma?: string;
-  sof_zman_shema_gra?: string;
-  sof_zman_tefila_ma?: string;
-  sof_zman_tefila_gra?: string;
-  chatzos?: string;
-  mincha_gedola?: string;
-  mincha_ketana?: string;
-  plag_mincha?: string;
-  sunset?: string;
-  tzeis_5_95_degrees?: string;
-  tzeis_8_5_degrees?: string;
-  tzeis_42_minutes?: string;
-  tzeis_72_minutes?: string;
-  chatzot_laila?: string;
-  astronomical_hour_ma?: string;
-  astronomical_hour_gra?: string;
+  alos: string | null;
+  sunrise: string | null;
+  misheyakir_10_2: string | null;
+  sof_zman_shema_ma: string | null;
+  sof_zman_shema_gra: string | null;
+  sof_zman_tefila_ma: string | null;
+  sof_zman_tefila_gra: string | null;
+  chatzos: string | null;
+  mincha_gedola: string | null;
+  mincha_ketana: string | null;
+  plag_mincha: string | null;
+  sunset: string | null;
+  tzeis_5_95_degrees: string | null;
+  tzeis_8_5_degrees: string | null;
+  tzeis_42_minutes: string | null;
+  tzeis_72_minutes: string | null;
+  chatzot_laila: string | null;
+  astronomical_hour_ma: string | null;
+  astronomical_hour_gra: string | null;
 }
