@@ -1,11 +1,9 @@
-import {Directive} from '@angular/core';
+import { Directive, HostBinding } from '@angular/core';
 
 @Directive({
   selector: '[appDefaultLayoutForm]',
-  // tslint:disable-next-line:no-host-metadata-property
-  host: {
-    '[style.min-width]': '"200px"',
-  }
 })
 export class DefaultLayoutFormDirective {
+  @HostBinding('style.min-width.rem')
+  minWidth = 20;
 }
