@@ -1,50 +1,77 @@
-import {Component, Inject, Injector, Input, OnDestroy, OnInit} from '@angular/core';
-import {TuiDialogService} from '@taiga-ui/core';
-import {PolymorpheusComponent} from '@tinkoff/ng-polymorpheus';
-import {DefaultLayoutMapComponent} from './default-layout-map/default-layout-map.component';
-import {Subscription} from 'rxjs';
-import {StoreService} from '@core/store';
+import { Component, Inject, Injector, OnDestroy, OnInit } from '@angular/core';
+import { TuiDialogService } from '@taiga-ui/core';
+import { PolymorpheusComponent } from '@tinkoff/ng-polymorpheus';
+import { DefaultLayoutMapComponent } from './default-layout-map/default-layout-map.component';
+import { Observable, Subscription } from 'rxjs';
+import { Select, Store } from '@ngxs/store';
+import {
+  AppState,
+  LocationModel,
+  LocationWithoutSourceModel,
+  SetLocationManually,
+} from '@core/state';
+import { TranslateService } from '@ngx-translate/core';
+import { map, shareReplay, switchMap } from 'rxjs/operators';
+import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout';
 
 @Component({
   selector: 'app-default-layout',
   templateUrl: './default-layout.component.html',
-  styleUrls: ['./default-layout.component.scss']
+  styleUrls: ['./default-layout.component.scss'],
 })
-export class DefaultLayoutComponent implements OnInit, OnDestroy {
-  cityName?: string;
+export class DefaultLayoutComponent implements OnDestroy {
+  @Select(AppState.location) location$!: Observable<LocationModel>;
+
+  readonly isHandset$: Observable<boolean> = this.breakpointObserver
+    .observe(Breakpoints.Handset)
+    .pipe(
+      map((result) => result.matches),
+      shareReplay(),
+    );
 
   private readonly onDestroy$: Subscription = new Subscription();
 
   constructor(
     @Inject(TuiDialogService) private readonly dialogService: TuiDialogService,
     @Inject(Injector) private readonly injector: Injector,
-    private readonly storeService: StoreService
-  ) {
-  }
-
-  ngOnInit(): void {
-    this.initCityName();
-  }
+    private readonly store: Store,
+    private readonly translateService: TranslateService,
+    private readonly breakpointObserver: BreakpointObserver,
+  ) {}
 
   ngOnDestroy(): void {
     this.onDestroy$.unsubscribe();
   }
 
-  openMap(): void {
-    this.onDestroy$.add(
-      this.dialogService.open(
-        new PolymorpheusComponent(DefaultLayoutMapComponent, this.injector),
-        {dismissible: true}
-      ).subscribe()
+  openMapDialog(): void {
+    const location: LocationModel | null = this.store.selectSnapshot(
+      AppState.location,
     );
-  }
+    if (!location) {
+      return;
+    }
 
-  private initCityName(): void {
     this.onDestroy$.add(
-      this.storeService.coords$
-        .subscribe(({cityName}) => {
-          this.cityName = cityName;
-        })
+      this.translateService
+        .get('default-layout.map-dialog-heading')
+        .pipe(
+          switchMap((heading) =>
+            this.dialogService.open<LocationWithoutSourceModel>(
+              new PolymorpheusComponent(
+                DefaultLayoutMapComponent,
+                this.injector,
+              ),
+              {
+                dismissible: true,
+                data: location,
+                label: heading,
+              },
+            ),
+          ),
+        )
+        .subscribe((location) => {
+          this.store.dispatch(new SetLocationManually(location));
+        }),
     );
   }
 }
