@@ -1,14 +1,11 @@
-import {NgModule} from '@angular/core';
-import {RouterModule, Routes} from '@angular/router';
-import {ZmanimComponent} from './zmanim.component';
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { ZmanimComponent } from './zmanim.component';
 
-const routes: Routes = [
-  {path: '', component: ZmanimComponent}
-];
+const routes: Routes = [{ path: '', component: ZmanimComponent }];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
+  exports: [RouterModule],
 })
-export class ZmanimRoutingModule {
-}
+export class ZmanimRoutingModule {}
