@@ -1,23 +1,22 @@
-import {NgModule} from '@angular/core';
-import {CommonModule} from '@angular/common';
-import {ZmanimRoutingModule} from './zmanim-routing.module';
-import {ZmanimComponent} from './zmanim.component';
-import {FormsModule, ReactiveFormsModule} from '@angular/forms';
-import {ZmanimFormComponent} from './zmanim-form/zmanim-form.component';
-import {ZmanimInfoComponent} from './zmanim-info/zmanim-info.component';
-import {DefaultLayoutModule} from '@shared/layout';
-import {TuiInputDateModule, TuiInputNumberModule} from '@taiga-ui/kit';
-import {TuiGroupModule} from '@taiga-ui/core';
-import {TranslateModule} from '@ngx-translate/core';
-import {TuiTableModule} from '@taiga-ui/addon-table';
-
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { ZmanimRoutingModule } from './zmanim-routing.module';
+import { ZmanimComponent } from './zmanim.component';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { ZmanimFormComponent } from './zmanim-form/zmanim-form.component';
+import { ZmanimInfoComponent } from './zmanim-info/zmanim-info.component';
+import { DefaultLayoutModule } from '@shared/layout';
+import {
+  TuiInputDateModule,
+  TuiInputNumberModule,
+  TuiIslandModule,
+} from '@taiga-ui/kit';
+import { TuiButtonModule, TuiGroupModule } from '@taiga-ui/core';
+import { TranslateModule } from '@ngx-translate/core';
+import { TuiTableModule } from '@taiga-ui/addon-table';
 
 @NgModule({
-  declarations: [
-    ZmanimComponent,
-    ZmanimFormComponent,
-    ZmanimInfoComponent,
-  ],
+  declarations: [ZmanimComponent, ZmanimFormComponent, ZmanimInfoComponent],
   imports: [
     CommonModule,
     FormsModule,
@@ -29,7 +28,8 @@ import {TuiTableModule} from '@taiga-ui/addon-table';
     TuiGroupModule,
     TranslateModule,
     TuiTableModule,
-  ]
+    TuiButtonModule,
+    TuiIslandModule,
+  ],
 })
-export class ZmanimModule {
-}
+export class ZmanimModule {}
